pyproject.toml
[project]
name = "digit-classifier"
version = "0.0.1"
description = "Digit Classifier uses an ML model (trained on the MNIST dataset) to classify a digit drawn by the user"
authors = [
    {name = "Helen Zhou"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = "=3.9.6"
dependencies = [
    "torch (<=2.2.2)",
    "torchvision (<=0.17.2)",
    "numpy (<2.0)",
    "torchmetrics (<=1.7.1)",
    "streamlit (<=1.45.0)",
    "streamlit-drawable-canvas (<=0.9.3)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
name = "digit-classifier"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
